+++ libs/components/colorpicker/projects/colorpicker/documentation/code-examples/colorpicker/programmatic/colorpicker-demo.module.ts
@@ -1,0 +1,34 @@
+import {
+  NgModule
+} from '@angular/core';
+
+import {
+  CommonModule
+} from '@angular/common';
+
+import {
+  FormsModule
+} from '@angular/forms';
+
+import {
+  SkyColorpickerModule
+} from 'projects/colorpicker/src/public-api';
+
+import {
+  ColorpickerDemoComponent
+} from './colorpicker-demo.component';
+
+@NgModule({
+  imports: [
+    CommonModule,
+    FormsModule,
+    SkyColorpickerModule
+  ],
+  declarations: [
+    ColorpickerDemoComponent
+  ],
+  exports: [
+    ColorpickerDemoComponent
+  ]
+})
+export class ColorpickerDemoModule { }